Massive Outfit Redesign and Reorganization
Re-balanced pretty much all weapons, as well as some other outfits.
Also used the chance to finally rename the images to follow the new "lore".
Changed capitalization and organization of sprites and effects to be more in line with vanilla.
Reorganized the sales definitions to make them better to look at and use (mostly alphabetical, but sorted by outfit type).