Updated Ships
Major balance changes for all ships, to be in line with the outfit re-balance. Reduced outfit space, weapon cap and engine cap, as well as some minor changes to shield/hull and cargo space.
Also updated most ship sprites with new post-processing and fixed all sprites to not be blurry anymore.