Updated the Map
Updated all systems in the central cluster with trade goods/prices and a basic set of fleets. Also rewrote some descriptions, adding a few new ones as well. Switched the systems of "Assuria" and "Curia", cut the link between "Sakura" and "Gargantua" and renamed the ringworld from "Avalon Station" to "Orbis Alteran", as it is once again a Quarg ringworld.