Ship recategorization
With the addition of Utility and Space Liner categories, some Alteran vessels could be moved to a better fitting category:
The Arqueus has become Utility vessel, as it serves both as a general civilian transport and a military vessel.
The Delicia can now be correctly categorized as a Space Liner.