Rework Map
Removed the western cluster, as it is not needed currently (no clear plan yet) and it would have needed a rework anyway.
Separated the other central and eastern clusters into different map files.
Renamed a handful of systems and planets.
Added attributes to all systems and planets currently in the central cluster.